curve_tool: fix log
